requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Frozen-Flask==1.0.2
itsdangerous==2.2.0
Jinja2==3.1.4
libsass==0.23.0
MarkupSafe==2.1.5
python-dotenv==1.0.1
Werkzeug==3.0.3
